Fixes #3766 - Introduce HTTP/2 API to batch frames.
Updates after review.
